requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
numpy>=1.26.2
pytest>=9.0.1

# Security Dependencies (RFC-PHOENIX-03 - Fase 3)
PyJWT>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0
PyYAML>=6.0.1
